multi-backend: create a backend and encapsulates multiple backends
This only implements and tests the basics of having a backend that encapsulates
trees into itself.

The basic idea is that the backend stores trees where some contained tree IDs are pointers into an inner backend.
Similarly, commits are also normal commits with pointers to commits from the inner backends.

Also extend the repo loading logic in preparation for broader integration.